Pin Alpine Linux version and install missing collectd plugin
The python:2.7-alpine image uses the latest major version of Alpine
Linux. When a new version of Alpine is released, changes in package
organization can break Dockerfile builds from one build of Via to the next.

In this case a "silent" update to Alpine broke collectd because the "disk"
plugin was split out into a separate package (see also [1]).

This commit pins the Alpine major version to 3.7, which is what we use
for "h" as well. This way we'll still get security updates to 3.7.x
packages but should avoid breaking changes.

[1] https://github.com/hypothesis/h/commit/7e8047db14f6804f8ca952623cf27e9f0ccd6bcc