Support customizing Hypothesis client config on a per-request basis
Support customizing aspects of the Hypothesis client configuration for a
given request by setting "via."-prefixed query params. These params are
stripped from the WSGI environment and are not seen by the proxy.

The initial use case is to support setting the "requestConfig" and
"openSidebar" client config properties, which will be needed by our LMS
integration.

In future we could also use this mechanism as a way to set "feature
flags" for requests in order to test out server-side behavior changes.