Unset {HTTP, HTTPS}_PROXY vars when running tests in Jenkins
These are set to point to squid when running the Docker container
normally, but interfere with dependency installation by pip.

In addition to being unset when running pip, they also now need to be
unset when running tox because the venv-update plugin may attempt to
connect to PyPI. This is a bit pointless in this context because the tox
env inside the Docker container always has up-to-date dependencies.