Add an ARIA label to the input field
Provide a label for the URL input. The placeholder text describes what the input
is for but that is only visible when the field is empty. For sighted
users we think this is adequate (cf. Google and Twitter search
homepages) but not for those using eg. screenreaders to navigate the
page.