Support running Via over SSL in development
Follow the convention in our other repositories of supporting running
the server over SSL locally if `.tlscert.pem` and `.tlskey.pem` files
are present.

uwsgi installed from PyPI does not support SSL, so instead add a
separate `make dev-ssl` command which uses gunicorn, which we already
use as the WSGI HTTP server in our other apps. This is easier than
trying to rebuild uwsgi with SSL support.

One small code change was required to support gunicorn, which is not to
rely on `REQUEST_URI` being present in the WSGI environ dict, as this is
uwsgi-specific.