Avoid injecting client into iframes - second attempt
The previous method relied on Via rewriting iframe `src` locations so that
the path started with `/if_/`. However there are some circumstances in
which this does not happen, such as when creating an iframe via
`document.createElement` and setting its `src` using
`iframe.setAttribute("src", "<location>")`. In those cases, the client
would be injected into both the iframe and the top-level Via frame.

This commit replaces that method with a more robust one which instead
relies on the fact that all proxied frames are served from the same
origin, Via's origin, and can therefore access each other's
`window.location`. Frames from different origins (such as the LMS app
embedding a Via iframe) will not be accessible. Therefore we can detect
whether the current frame is the top-most Via frame (as opposed to an
iframe proxied through Via) and only inject the client in that case.

In the context of the LMS app, if the client was injected into a
document proxied through Via as well as iframes in that document proxied
through Via, then multiple instances of the sidebar app would be created
with different client configurations. The main frame would be configured
to fetch client configuration from the LMS app but child iframes would
use the client's default configuration, i.e. it would fetch annotations
and account info from the first-party authority in h, instead of the lms
app's authority. Due to the client's problematic process for
discovering content frames [1] the clients in iframes would end up
connecting to the main frame and fetching annotations for it from the
first-party authority, with the result that annotations from the
"Public" group would end up showing up in the main page even if if the
sidebar in the main page was showing annotations from a third-party
group, such as a group created by the LMS app for the current course.

Fixes https://github.com/hypothesis/via/issues/119

[1] See https://github.com/hypothesis/client/issues/187 and
    https://github.com/hypothesis/client/issues/249