Run uwsgi with newrelic monitoring
In order to do this, the [New Relic
docs](https://docs.newrelic.com/docs/agents/python-agent/web-frameworks-servers/python-agent-uwsgi-web-server)
suggest running with the following uwsgi settings:
- enable-threads
    Allows the the creation of background threads from Python. This is a
required option for New Relic as this is how it collects and sends data.
- single-interpreter
   Limits the number of web applications to one per process. This is
recomended as a safety precaution.

Since we are running on a newer version of uwsgi and out of concern for
unforeseen consequences, I have chosen to *not* enable single-interpreter.