Point New Relic to the main app entry point
New Relic does not have built in middleware integation
with uwsgi so we need to point it to the main uwsgi entry
point of the application. Override the transaction name to proxy
to prevent New Relic from setting the transaction name equal
to whatever page via is proxying and ending up with thousands of
different transaction names. request.uri will give us the webpage
that via is proxying.