Add scripts to update PDF.js from upstream
Add a script which downloads the latest version of PDF.js and the
default viewer application from GitHub and regenerates the viewer HTML
template that Via serves.

These scripts are intended to make it much easier to update PDF.js in
future.

In the process the custom code used to load the Hypothesis client into the
viewer was updated:

 - Extract the code into a separate `pdfjs-init.js` script so that it is
   more obviously separate from the generated code.
 - Modify the code to work with PDF.js v2
 - Try to use only public APIs to load the document, to make future
   PDF.js updates easier