Avoid trying to load the PDF until the PDF viewer is fully initialized
When `app.open` was called before the PDF viewer had fully initialized,
the PDF sometimes failed to load with a "this.toolbar is null" exception
being reported. From tracing the code, it looks like what was happening
in that case is that `app.open` caused a call to `PDFViewerApplication.load`
to happen before all of the properties of `PDFViewerApplication` had been
fully initialized.

Fixes the issue described in https://github.com/hypothesis/via/issues/81#issuecomment-530800859