Replace HTML snippet injected into proxied pages
Replace the default HTML snippet that pywb injects at the top of proxied
pages with one which just adds the the banner that loads the Hypothesis
client and adds a `<link rel="canonical">`, but does not inject the
scripts that monkey-patch DOM APIs to do rewrite URL references and
route XHR requests through our servers.

This is part of an experiment into determining whether a proxy service
can be far more minimal and still be useful.