Remove ngs-httpclient dependency
This dependency was originally added to support sites that require SNI [1]
but the sites mentioned in the issue and the test page at
https://check-tls.akamaized.net seem to work without it.

This dependency in turn resulted in Via having a bunch of outdated
dependencies (eg. cryptography) which appears to be a cause of a startup
crash when running the dev server on the latest version of macOS
(the issue I encountered matched https://stackoverflow.com/questions/58563849/)

If this approach doesn't work and it turns out that ngs-httpclient is
still required, we may have to update its transitive dependencies
instead.

[1] https://github.com/hypothesis/via/issues/8