Install python2 and py2-pip as runtime rather than build-only dependencies
Python 2 used to be installed by supervisor, but that is no longer the
case with Alpine 3.11.

Via depends on the `pkg_resources` module which is part of `setuptools`,
which in turn is included with pip. If py2-pip is installed and then
removed as part of installing build-only dependencies, it ends up
removing the `setuptools` package.