Support specifying blocklist via external file
Instead of hardcoding the blocklist in Python code, define it in a
simple text-based file format which the Via Python app reads from a
local file path specified via the `BLOCKLIST_PATH` environment variable.

Within the Docker container, a shell-script will fetch the blocklist
from a URL every N seconds and write that to a file which the Python
worker processes will then read.