If merged, this commit will remove the goreplay service from the via stack. (#219)
Goreplay is a service which listens for traffic to your app and uses it as
input to another app which is being tested. It was added to via long ago
in order to test a via replacement. It is now hitting a bug and spamming
error logs with stack traces. This could be affecting performance and
might even present a security issue depending on the nature of the crash.